Change with_capacity constructor to also be a convenience method
Meaning that it will create a Stash that uses usize as index type.
In the general case where a custom index type might be needed the user
can construct the Stash with Stash::default() and then call reserve() to
allocate needed capacity. This way we don't break existing code and the
new use case of custom index types is still possible (and also pretty
convenient.. not bad.)

Also add a new example to show how a custom index type can be constructed
and used.